refactor: move parseSchemaFileInfo() to db package
