+++ server/webhook.go
@@ -12,7 +12,6 @@ import (
 	"io"
 	"net/http"
 	"path"
-	"regexp"
 	"strings"
 
 	"github.com/labstack/echo/v4"
@@ -218,6 +217,11 @@ func (s *Server) processPushEvent(ctx context.Context, repositoryList []*api.Rep
 
 	var createdMessages []string
 	for _, item := range distinctFileList {
+		log.Debug("Processing file",
+			zap.String("file", item.FileName),
+			zap.String("commit", item.Commit.ID),
+		)
+
 		var createdMessageList []string
 		repoID2ActivityCreateList := make(map[int][]*api.ActivityCreate)
 		for _, repo := range repositoryList {
@@ -287,11 +291,6 @@ func (s *Server) processFile(ctx context.Context, pushEvent *vcs.PushEvent, repo
 		}
 	}
 
-	log.Debug("Processing file",
-		zap.String("file", file),
-		zap.String("commit", pushEvent.FileCommit.ID),
-	)
-
 	if !strings.HasPrefix(file, repo.BaseDirectory) {
 		log.Debug("Ignored file outside the base directory",
 			zap.String("file", file),
@@ -300,7 +299,7 @@ func (s *Server) processFile(ctx context.Context, pushEvent *vcs.PushEvent, repo
 		return "", false, nil, nil
 	}
 
-	schemaInfo, err := parseSchemaFileInfo(repo.BaseDirectory, repo.SchemaPathTemplate, file)
+	schemaInfo, err := db.ParseSchemaFileInfo(repo.BaseDirectory, repo.SchemaPathTemplate, file)
 	if err != nil {
 		log.Debug("Failed to parse schema file info",
 			zap.String("file", file),
@@ -557,8 +556,8 @@ func (s *Server) readFileContent(ctx context.Context, pushEvent *vcs.PushEvent, 
 
 // prepareIssueFromPushEventSDL returns the migration info and a list of update
 // schema details derived from the given push event for SDL.
-func (s *Server) prepareIssueFromPushEventSDL(ctx context.Context, repo *api.Repository, pushEvent *vcs.PushEvent, schemaInfo map[string]string, file string) ([]*api.MigrationDetail, []*api.ActivityCreate) {
-	dbName := schemaInfo["DB_NAME"]
+func (s *Server) prepareIssueFromPushEventSDL(ctx context.Context, repo *api.Repository, pushEvent *vcs.PushEvent, schemaInfo *db.MigrationInfo, file string) ([]*api.MigrationDetail, []*api.ActivityCreate) {
+	dbName := schemaInfo.Database
 	if dbName == "" {
 		log.Debug("Ignored schema file without a database name", zap.String("file", file))
 		return nil, nil
@@ -571,7 +570,7 @@ func (s *Server) prepareIssueFromPushEventSDL(ctx context.Context, repo *api.Rep
 	}
 
 	activityCreateList := []*api.ActivityCreate{}
-	envName := schemaInfo["ENV_NAME"]
+	envName := schemaInfo.Environment
 	var migrationDetailList []*api.MigrationDetail
 	if repo.Project.TenantMode == api.TenantModeTenant {
 		migrationDetailList = append(migrationDetailList,
@@ -697,46 +696,6 @@ func (s *Server) tryUpdateTasksFromModifiedFile(ctx context.Context, databases [
 		}
 	}
 	return nil
-}
-
-// parseSchemaFileInfo attempts to parse the given schema file path to extract
-// the schema file info. It returns (nil, nil) if it doesn't looks like a schema
-// file path.
-//
-// The possible keys for the returned map are: "ENV_NAME", "DB_NAME".
-func parseSchemaFileInfo(baseDirectory, schemaPathTemplate, file string) (map[string]string, error) {
-	if schemaPathTemplate == "" {
-		return nil, nil
-	}
-
-	// Escape "." characters to match literals instead of using it as a wildcard.
-	schemaFilePathRegex := strings.ReplaceAll(schemaPathTemplate, ".", `\.`)
-
-	placeholders := []string{
-		"ENV_NAME",
-		"DB_NAME",
-	}
-	for _, placeholder := range placeholders {
-		schemaFilePathRegex = strings.ReplaceAll(schemaFilePathRegex, fmt.Sprintf("{{%s}}", placeholder), fmt.Sprintf("(?P<%s>[a-zA-Z0-9+-=/_#?!$. ]+)", placeholder))
-	}
-
-	// NOTE: We do not want to use filepath.Join here because we always need "/" as the path separator.
-	re, err := regexp.Compile(path.Join(baseDirectory, schemaFilePathRegex))
-	if err != nil {
-		return nil, errors.Wrap(err, "compile schema file path regex")
-	}
-	match := re.FindStringSubmatch(file)
-	if len(match) == 0 {
-		return nil, nil
-	}
-
-	info := make(map[string]string)
-	// Skip the first item because it is always the empty string, see docstring of
-	// the SubexpNames() method.
-	for i, name := range re.SubexpNames()[1:] {
-		info[name] = match[i+1]
-	}
-	return info, nil
 }
 
 // computeDatabaseSchemaDiff computes the diff between current database schema
