+++ server/webhook_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestValidateGitHubWebhookSignature256(t *testing.T) {
@@ -76,67 +75,5 @@ func TestParseBranchNameFromGitHubRefs(t *testing.T) {
 		} else {
 			assert.Equal(t, test.expect, branch)
 		}
-	}
-}
-
-func TestParseSchemaFileInfo(t *testing.T) {
-	tests := []struct {
-		name               string
-		baseDirectory      string
-		schemaPathTemplate string
-		file               string
-		schemaInfo         map[string]string
-	}{
-		{
-			name:               "no schemaPathTemplate",
-			baseDirectory:      "",
-			schemaPathTemplate: "",
-			file:               "Test/testdb__LATEST.sql",
-			schemaInfo:         nil,
-		},
-		{
-			name:               "only has DB_NAME",
-			baseDirectory:      "",
-			schemaPathTemplate: "{{DB_NAME}}__LATEST.sql",
-			file:               "testdb__LATEST.sql",
-			schemaInfo: map[string]string{
-				"DB_NAME": "testdb",
-			},
-		},
-		{
-			name:               "has both ENV_NAME and DB_NAME",
-			baseDirectory:      "",
-			schemaPathTemplate: "{{ENV_NAME}}/{{DB_NAME}}__LATEST.sql",
-			file:               "Test/testdb__LATEST.sql",
-			schemaInfo: map[string]string{
-				"ENV_NAME": "Test",
-				"DB_NAME":  "testdb",
-			},
-		},
-
-		{
-			name:               "baseDirectory does not match",
-			baseDirectory:      "bytebase",
-			schemaPathTemplate: "{{ENV_NAME}}/{{DB_NAME}}__LATEST.sql",
-			file:               "Test/testdb__LATEST.sql",
-			schemaInfo:         nil,
-		},
-		{
-			name:               "baseDirectory with both ENV_NAME and DB_NAME",
-			baseDirectory:      "bytebase",
-			schemaPathTemplate: "{{ENV_NAME}}/{{DB_NAME}}__LATEST.sql",
-			file:               "bytebase/Test/testdb__LATEST.sql",
-			schemaInfo: map[string]string{
-				"ENV_NAME": "Test",
-				"DB_NAME":  "testdb",
-			},
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			got, err := parseSchemaFileInfo(test.baseDirectory, test.schemaPathTemplate, test.file)
-			require.NoError(t, err)
-			assert.Equal(t, test.schemaInfo, got)
-		})
 	}
 }
